+++ client/idrepo/enduser/src/main/java/org/apache/syncope/client/enduser/rest/AbstractAnyRestClient.java
@@ -27,7 +27,7 @@ public abstract class AbstractAnyRestClient<TO extends AnyTO> extends BaseRestCl
 
     protected abstract Class<? extends AnyService<TO>> getAnyServiceClass();
 
-    public abstract int count(String realm, String fiql, String type);
+    public abstract long count(String realm, String fiql, String type);
 
     public TO read(final String key) {
         return getService(getAnyServiceClass()).read(key);
